+++ src/greynoise/api/__init__.py
@@ -25,7 +25,7 @@ def initialize_cache(cache_max_size, cache_ttl):
     return cache
 
 
-class GreyNoise(object):
+class GreyNoise(object):  # pylint: disable=R0205,R0902
 
     """GreyNoise API client.
 
@@ -45,6 +45,7 @@ class GreyNoise(object):
     EP_INTERESTING = "interesting/{ip_address}"
     EP_NOISE_MULTI = "noise/multi/quick"
     EP_NOISE_CONTEXT = "noise/context/{ip_address}"
+    EP_NOISE_CONTEXT_MULTI = "noise/multi/context"
     EP_COMMUNITY_IP = "v3/community/{ip_address}"
     EP_META_METADATA = "meta/metadata"
     EP_PING = "ping"
@@ -97,7 +98,7 @@ class GreyNoise(object):
         cache_max_size=None,
         cache_ttl=None,
         offering=None,
-    ):
+    ):  # pylint: disable=R0913
         if any(
             configuration_value is None
             for configuration_value in (api_key, timeout, api_server, proxy, offering)
@@ -153,9 +154,11 @@ class GreyNoise(object):
         if params is None:
             params = {}
 
-        user_agent_parts = ["GreyNoise/{}".format(__version__)]
+        user_agent_parts = ["GreyNoise/{}".format(__version__)]  # pylint: disable=C0209
         if self.integration_name:
-            user_agent_parts.append("({})".format(self.integration_name))
+            user_agent_parts.append(
+                "({})".format(self.integration_name)
+            )  # pylint: disable=C0209
         headers = {
             "User-Agent": " ".join(user_agent_parts),
             "key": self.api_key,
@@ -224,10 +227,11 @@ class GreyNoise(object):
             response = [
                 {"message": "Quick Lookup not supported with Community offering"}
             ]
-            return response
         else:
             analyzer = Analyzer(self)
-            return analyzer.analyze(text)
+            response = analyzer.analyze(text)
+
+        return response
 
     def filter(self, text, noise_only=False, riot_only=False):
         """Filter lines that contain IP addresses from a given text.
@@ -246,8 +250,8 @@ class GreyNoise(object):
         :rtype: iterable
 
         """
-        filter = Filter(self)
-        for filtered_chunk in filter.filter(
+        gnfilter = Filter(self)
+        for filtered_chunk in gnfilter.filter(
             text, noise_only=noise_only, riot_only=riot_only
         ):
             yield filtered_chunk
@@ -263,7 +267,6 @@ class GreyNoise(object):
             response = {
                 "message": "Interesting report not supported with Community offering"
             }
-            return response
         else:
             LOGGER.debug(
                 "Reporting interesting IP: %s...", ip_address, ip_address=ip_address
@@ -272,9 +275,10 @@ class GreyNoise(object):
 
             endpoint = self.EP_INTERESTING.format(ip_address=ip_address)
             response = self._request(endpoint, method="post")
-            return response
 
-    def ip(self, ip_address):
+        return response
+
+    def ip(self, ip_address):  # pylint: disable=C0103
         """Get context associated with an IP address.
 
         :param ip_address: IP address to use in the look-up.
@@ -320,7 +324,6 @@ class GreyNoise(object):
         """Run GNQL query."""
         if self.offering == "community":
             response = {"message": "GNQL not supported with Community offering"}
-            return response
         else:
             LOGGER.debug(
                 "Running GNQL query: %s...",
@@ -335,9 +338,10 @@ class GreyNoise(object):
             if scroll is not None:
                 params["scroll"] = scroll
             response = self._request(self.EP_GNQL, params=params)
-            return response
 
-    def quick(self, ip_addresses, include_invalid=False):  # noqa: C901
+        return response
+
+    def quick(self, ip_addresses, include_invalid=False):  # pylint: disable=R0912,R0914
         """Get activity associated with one or more IP addresses.
 
         :param ip_addresses: One or more IP addresses to use in the look-up.
@@ -353,7 +357,6 @@ class GreyNoise(object):
             response = [
                 {"message": "Quick Lookup not supported with Community offering"}
             ]
-            return response
         else:
             if isinstance(ip_addresses, str):
                 ip_addresses = ip_addresses.split(",")
@@ -363,7 +366,7 @@ class GreyNoise(object):
             valid_ip_addresses = [
                 ip_address
                 for ip_address in ip_addresses
-                if validate_ip(ip_address, strict=False)
+                if validate_ip(ip_address, strict=False, print_warning=False)
             ]
 
             if self.use_cache:
@@ -411,31 +414,132 @@ class GreyNoise(object):
                     else:
                         results.append(result)
 
-            [
-                results.append({"ip": ip, "noise": False, "code": "404"})
-                for ip in ip_addresses
-                if ip not in valid_ip_addresses and include_invalid
-            ]
+            if include_invalid:
+                for ip_address in ip_addresses:
+                    if ip_address not in valid_ip_addresses:
+                        results.append(
+                            {
+                                "ip": ip_address,
+                                "noise": False,
+                                "riot": False,
+                                "code": "404",
+                            }
+                        )
 
             for result in results:
                 code = result["code"]
                 result["code_message"] = self.CODE_MESSAGES.get(
                     code, self.UNKNOWN_CODE_MESSAGE.format(code)
                 )
-            return results
+            response = results
+
+        return response
+
+    def ip_multi(self, ip_addresses, include_invalid=False):  # pylint: disable=R0912
+        """Get activity associated with one or more IP addresses.
+
+        :param ip_addresses: One or more IP addresses to use in the look-up.
+        :type ip_addresses: str | list
+        :return: Bulk status information for IP addresses.
+        :rtype: dict
+
+        :param include_invalid: True or False
+        :type include_invalid: bool
+
+        """
+        if self.offering == "community":  # pylint: disable=R1702
+            results = [
+                {"message": "IP Multi Lookup not supported with Community offering"}
+            ]
+        else:
+            if isinstance(ip_addresses, str):
+                ip_addresses = ip_addresses.split(",")
+
+            LOGGER.debug("Getting noise context...", ip_addresses=ip_addresses)
+
+            valid_ip_addresses = [
+                ip_address
+                for ip_address in ip_addresses
+                if validate_ip(ip_address, strict=False, print_warning=False)
+            ]
+
+            if self.use_cache:
+                cache = self.ip_context_cache
+                # Keep the same ordering as in the input
+                ordered_results = OrderedDict(
+                    (ip_address, cache.get(ip_address))
+                    for ip_address in valid_ip_addresses
+                )
+                api_ip_addresses = [
+                    ip_address
+                    for ip_address, result in ordered_results.items()
+                    if result is None
+                ]
+                if api_ip_addresses:
+                    api_results = []
+                    chunks = more_itertools.chunked(
+                        api_ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
+                    )
+                    for chunk in chunks:
+                        api_result = self._request(
+                            self.EP_NOISE_CONTEXT_MULTI,
+                            method="post",
+                            json={"ips": chunk},
+                        )
+
+                        api_result = api_result["data"]
+
+                        if isinstance(api_result, list):
+                            api_results.extend(api_result)
+                        else:
+                            api_results.append(api_result)
+
+                        for ip_address in valid_ip_addresses:
+                            if ip_address not in api_results:
+                                api_results.append({"ip": ip_address, "seen": False})
+
+                    for result in api_results:
+                        ip_address = result["ip"]
+
+                        ordered_results[ip_address] = cache.setdefault(
+                            ip_address, result
+                        )
+
+                results = list(ordered_results.values())
+
+            else:
+                results = []
+                chunks = more_itertools.chunked(
+                    valid_ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
+                )
+                for chunk in chunks:
+                    result = self._request(
+                        self.EP_NOISE_CONTEXT_MULTI, json={"ips": chunk}
+                    )
+                    if isinstance(result, list):
+                        results.extend(result)
+                    else:
+                        results.append(result)
+
+            if include_invalid:
+                for ip_address in ip_addresses:
+                    if ip_address not in valid_ip_addresses:
+                        results.append({"ip": ip_address, "seen": False})
+
+        return results
 
     def stats(self, query, count=None):
         """Run GNQL stats query."""
         if self.offering == "community":
             response = {"message": "Stats Query not supported with Community offering"}
-            return response
         else:
             LOGGER.debug("Running GNQL stats query: %s...", query, query=query)
             params = {"query": query}
             if count is not None:
                 params["count"] = count
             response = self._request(self.EP_GNQL_STATS, params=params)
-            return response
+
+        return response
 
     def metadata(self):
         """Get metadata."""
@@ -443,11 +547,11 @@ class GreyNoise(object):
             response = {
                 "message": "Metadata lookup not supported with Community offering"
             }
-            return response
         else:
             LOGGER.debug("Getting metadata...")
             response = self._request(self.EP_META_METADATA)
-            return response
+
+        return response
 
     def test_connection(self):
         """Test the API connection and API key."""
@@ -466,7 +570,6 @@ class GreyNoise(object):
         """
         if self.offering == "community":
             response = {"message": "RIOT lookup not supported with Community offering"}
-            return response
         else:
             LOGGER.debug("Checking RIOT for %s...", ip_address, ip_address=ip_address)
             validate_ip(ip_address)
@@ -477,4 +580,4 @@ class GreyNoise(object):
             if "ip" not in response:
                 response["ip"] = ip_address
 
-            return response
+        return response