+++ tests/cli/test_subcommand.py
@@ -81,7 +81,7 @@ class TestAnalyze(object):
         },
     }
     DEFAULT_OUTPUT = textwrap.dedent(
-        u"""\
+        """\
         ╔═══════════════════════════╗
         ║          Analyze          ║
         ╚═══════════════════════════╝
@@ -664,7 +664,7 @@ class TestQuery(object):
         result = runner.invoke(subcommand.query, ["-f", "json", query])
         assert result.exit_code == 0
         assert result.output.strip("\n") == expected
-        api_client.query.assert_called_with(query=query)
+        api_client.query.assert_called_with(query=query, size=None, scroll=None)
 
     def test_input_file(self, api_client):
         """Run query from input file."""
@@ -677,7 +677,7 @@ class TestQuery(object):
         result = runner.invoke(subcommand.query, ["-f", "json", "-i", StringIO(query)])
         assert result.exit_code == 0
         assert result.output.strip("\n") == expected
-        api_client.query.assert_called_with(query=query)
+        api_client.query.assert_called_with(query=query, size=None, scroll=None)
 
     def test_stdin_input(self, api_client):
         """Run query from stdin."""
@@ -690,7 +690,7 @@ class TestQuery(object):
         result = runner.invoke(subcommand.query, ["-f", "json"], input=query)
         assert result.exit_code == 0
         assert result.output.strip("\n") == expected
-        api_client.query.assert_called_with(query=query)
+        api_client.query.assert_called_with(query=query, size=None, scroll=None)
 
     def test_no_query_passed(self, api_client):
         """Usage is returned if no query or input file is passed."""
@@ -926,6 +926,191 @@ class TestQuick(object):
             load_config.return_value = {"api_key": ""}
             result = runner.invoke(
                 subcommand.quick,
+                ["8.8.8.8"],
+                parent=Context(main, info_name="greynoise"),
+            )
+            assert result.exit_code == -1
+            assert "Error: API key not found" in result.output
+
+
+class TestIPMulti(object):
+    """IP-Multi subcommand tests."""
+
+    @pytest.mark.parametrize(
+        "ip_address, output_format, expected",
+        (
+            (
+                "8.8.8.8",
+                "json",
+                json.dumps(
+                    [{"ip": "8.8.8.8", "noise": True}], indent=4, sort_keys=True
+                ),
+            ),
+            (
+                "8.8.8.8",
+                "xml",
+                textwrap.dedent(
+                    """\
+                    <?xml version="1.0" ?>
+                    <root>
+                    \t<item>
+                    \t\t<ip>8.8.8.8</ip>
+                    \t\t<noise>True</noise>
+                    \t</item>
+                    </root>"""
+                ),
+            ),
+            (
+                "8.8.8.8",
+                "txt",
+                "8.8.8.8 is classified as NOT NOISE.",
+            ),
+        ),
+    )
+    def test_ip_multi(self, api_client, ip_address, output_format, expected):
+        """Quickly check IP address."""
+        runner = CliRunner()
+
+        api_client.ip_multi.return_value = [
+            OrderedDict((("ip", ip_address), ("noise", True)))
+        ]
+
+        result = runner.invoke(subcommand.ip_multi, ["-f", output_format, ip_address])
+        assert result.exit_code == 0
+        assert result.output.strip("\n") == expected
+        api_client.ip_multi.assert_called_with(ip_addresses=[ip_address])
+
+    @pytest.mark.parametrize(
+        "ip_addresses, mock_response, expected",
+        (
+            (
+                ["8.8.8.8", "8.8.8.9"],
+                [
+                    OrderedDict([("ip", "8.8.8.8"), ("noise", True)]),
+                    OrderedDict([("ip", "8.8.8.9"), ("noise", False)]),
+                ],
+                json.dumps(
+                    [
+                        {"ip": "8.8.8.8", "noise": True},
+                        {"ip": "8.8.8.9", "noise": False},
+                    ],
+                    indent=4,
+                    sort_keys=True,
+                ),
+            ),
+        ),
+    )
+    def test_input_file(self, api_client, ip_addresses, mock_response, expected):
+        """Quickly check IP address from input file."""
+        runner = CliRunner()
+
+        api_client.ip_multi.return_value = mock_response
+
+        result = runner.invoke(
+            subcommand.ip_multi, ["-f", "json", "-i", StringIO("\n".join(ip_addresses))]
+        )
+        assert result.exit_code == 0
+        assert result.output.strip("\n") == expected
+        api_client.ip_multi.assert_called_with(ip_addresses=ip_addresses)
+
+    @pytest.mark.parametrize(
+        "ip_addresses, mock_response, expected",
+        (
+            (
+                ["8.8.8.8", "8.8.8.9"],
+                [
+                    OrderedDict([("ip", "8.8.8.8"), ("noise", True)]),
+                    OrderedDict([("ip", "8.8.8.9"), ("noise", False)]),
+                ],
+                json.dumps(
+                    [
+                        {"ip": "8.8.8.8", "noise": True},
+                        {"ip": "8.8.8.9", "noise": False},
+                    ],
+                    indent=4,
+                    sort_keys=True,
+                ),
+            ),
+        ),
+    )
+    def test_stdin_input(self, api_client, ip_addresses, mock_response, expected):
+        """Quickly check IP address from stdin."""
+        runner = CliRunner()
+
+        api_client.ip_multi.return_value = mock_response
+
+        result = runner.invoke(
+            subcommand.ip_multi, ["-f", "json"], input="\n".join(ip_addresses)
+        )
+        assert result.exit_code == 0
+        assert result.output.strip("\n") == expected
+        api_client.ip_multi.assert_called_with(ip_addresses=ip_addresses)
+
+    def test_no_ip_address_passed(self, api_client):
+        """Usage is returned if no IP address or input file is passed."""
+        runner = CliRunner()
+
+        with patch("greynoise.cli.helper.sys") as sys:
+            sys.stdin.isatty.return_value = True
+            result = runner.invoke(
+                subcommand.ip_multi, parent=Context(main, info_name="greynoise")
+            )
+        assert result.exit_code == -1
+        assert "Usage: greynoise ip-multi" in result.output
+        api_client.ip_multi.assert_not_called()
+
+    def test_input_file_invalid_ip_addresses_passed(self, api_client):
+        """Error returned if only invalid IP addresses are passed in input file."""
+        runner = CliRunner()
+
+        expected = (
+            "Error: at least one valid IP address must be passed either as an "
+            "argument (IP_ADDRESS) or through the -i/--input_file option."
+        )
+
+        result = runner.invoke(
+            subcommand.ip_multi,
+            ["-i", StringIO("not-an-ip")],
+            parent=Context(main, info_name="greynoise"),
+        )
+        assert result.exit_code == -1
+        assert "Usage: greynoise ip-multi" in result.output
+        assert expected in result.output
+        api_client.ip_multi.assert_not_called()
+
+    def test_invalid_ip_address_as_argument(self, api_client):
+        """Quick subcommand fails when ip_address is invalid."""
+        runner = CliRunner()
+
+        expected = "Error: Invalid value for '[IP_ADDRESS]...': not-an-ip\n"
+
+        result = runner.invoke(subcommand.ip_multi, ["not-an-ip"])
+        assert result.exit_code == 2
+        assert "Usage: ip-multi [OPTIONS] [IP_ADDRESS]..." in result.output
+        assert expected in result.output
+        api_client.ip_multi.assert_not_called()
+
+    def test_request_failure(self, api_client):
+        """Error is displayed on API request failure."""
+        runner = CliRunner()
+
+        api_client.ip_multi.side_effect = RequestFailure(
+            401, {"message": "forbidden", "status": "error"}
+        )
+        expected = "API error: forbidden"
+
+        result = runner.invoke(subcommand.ip_multi, ["8.8.8.8"])
+        assert result.exit_code == -1
+        assert expected in result.output
+
+    def test_api_key_not_found(self):
+        """Error is displayed if API key is not found."""
+        runner = CliRunner()
+
+        with patch("greynoise.cli.decorator.load_config") as load_config:
+            load_config.return_value = {"api_key": ""}
+            result = runner.invoke(
+                subcommand.ip_multi,
                 ["8.8.8.8"],
                 parent=Context(main, info_name="greynoise"),
             )